requirements.txt
flask
pystray
pillow 